+++ src/js/control-bar/volume-control/volume-bar.js
@@ -4,7 +4,6 @@
 import Slider from '../../slider/slider.js';
 import Component from '../../component.js';
 import * as Fn from '../../utils/fn.js';
-import roundFloat from '../../utils/round-float.js';
 
 // Required children
 import VolumeLevel from './volume-level.js';
@@ -90,8 +89,9 @@ class VolumeBar extends Slider {
    */
   updateARIAAttributes() {
     // Current value of volume bar as a percentage
-    this.el_.setAttribute('aria-valuenow', roundFloat(this.player_.volume()*100, 2));
-    this.el_.setAttribute('aria-valuetext', roundFloat(this.player_.volume()*100, 2)+'%');
+    let volume = (this.player_.volume() * 100).toFixed(2);
+    this.el_.setAttribute('aria-valuenow', volume);
+    this.el_.setAttribute('aria-valuetext', volume + '%');
   }
 
 }