+++ src/js/tech/tech.js
@@ -538,6 +538,17 @@ Tech.withSourceHandlers = function(_Tech){
     return '';
   };
 
+  let originalSeekable = _Tech.prototype.seekable;
+
+  // when a source handler is registered, prefer its implementation of
+  // seekable when present.
+  _Tech.prototype.seekable = function() {
+    if (this.sourceHandler_ && this.sourceHandler_.seekable) {
+      return this.sourceHandler_.seekable();
+    }
+    return originalSeekable.call(this);
+  };
+
    /*
     * Create a function for setting the source using a source object
     * and source handlers.
@@ -566,16 +577,6 @@ Tech.withSourceHandlers = function(_Tech){
     this.sourceHandler_ = sh.handleSource(source, this);
     this.on('dispose', this.disposeSourceHandler);
 
-    this.originalSeekable_ = this.seekable;
-    // when a source handler is registered, prefer its implementation of
-    // seekable when present.
-    this.seekable = function() {
-      if (this.sourceHandler_ && this.sourceHandler_.seekable) {
-        return this.sourceHandler_.seekable();
-      }
-      return this.originalSeekable_.call(this);
-    };
-
     return this;
   };
 
@@ -585,7 +586,6 @@ Tech.withSourceHandlers = function(_Tech){
    _Tech.prototype.disposeSourceHandler = function(){
     if (this.sourceHandler_ && this.sourceHandler_.dispose) {
       this.sourceHandler_.dispose();
-      this.seekable = this.originalSeekable_;
     }
   };