+++ CHANGELOG.md
@@ -81,6 +81,7 @@ CHANGELOG
 * @misteroneill restore some properties on window.videojs. ([view](https://github.com/videojs/video.js/pull/2395))
 * @misteroneill restore some 4.x utilities and remove deprecated functionality ([view](https://github.com/videojs/video.js/pull/2406))
 * @heff use a synchronous ready() internally ([view](https://github.com/videojs/video.js/pull/2392))
+* @nickygerritsen scrubbing() is a method, not a property ([view](https://github.com/videojs/video.js/pull/2411))
 
 --------------------