+++ CHANGELOG.md
@@ -83,6 +83,7 @@ CHANGELOG
 * @heff use a synchronous ready() internally ([view](https://github.com/videojs/video.js/pull/2392))
 * @nickygerritsen scrubbing() is a method, not a property ([view](https://github.com/videojs/video.js/pull/2411))
 * @sirlancelot change &quot;video&quot; to &quot;media&quot; in error messages ([view](https://github.com/videojs/video.js/pull/2409))
+* @sirlancelot change &quot;video&quot; to &quot;media&quot; in error messages ([view](https://github.com/videojs/video.js/pull/2409))
 
 --------------------