+++ CHANGELOG.md
@@ -89,6 +89,7 @@ CHANGELOG
 * @misteroneill restore Html5.Events ([view](https://github.com/videojs/video.js/pull/2421))
 * @misteroneill removed the deprecated Component init method ([view](https://github.com/videojs/video.js/pull/2427))
 * @misteroneill restore videojs.formatTime ([view](https://github.com/videojs/video.js/pull/2420))
+* @misteroneill include child components with &#x60;true&#x60; in options ([view](https://github.com/videojs/video.js/pull/2424))
 
 --------------------