+++ CHANGELOG.md
@@ -90,6 +90,7 @@ CHANGELOG
 * @misteroneill removed the deprecated Component init method ([view](https://github.com/videojs/video.js/pull/2427))
 * @misteroneill restore videojs.formatTime ([view](https://github.com/videojs/video.js/pull/2420))
 * @misteroneill include child components with &#x60;true&#x60; in options ([view](https://github.com/videojs/video.js/pull/2424))
+* @misteroneill create video.novtt.js in dist builds ([view](https://github.com/videojs/video.js/pull/2447))
 
 --------------------